backend/requirements.txt
beanie==1.26.0
fastapi==0.111.0
motor==3.4.0
pydantic==2.7.4
python-dotenv==1.0.1
uvicorn==0.30.1
bcrypt==4.1.3
PyJWT==2.8.0
fastapi-mail==1.4.1
pytest==8.2.2